+++ src/kiwano/2d/FrameSequence.cpp
@@ -18,9 +18,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "FrameSequence.h"
-#include "Frame.h"
-#include "../base/Logger.h"
+#include <kiwano/2d/FrameSequence.h>
+#include <kiwano/2d/Frame.h>
+#include <kiwano/base/Logger.h>
 
 namespace kiwano
 {
@@ -59,7 +59,7 @@ namespace kiwano
 		}
 	}
 
-	FramePtr FrameSequence::GetFrame(UInt32 index) const
+	FramePtr FrameSequence::GetFrame(size_t index) const
 	{
 		KGE_ASSERT(index < frames_.size());
 		return frames_[index];
